# public/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'API Documentation'
  description: "TinoDeploy's API documentation."
  version: 1.0.0
servers:
  -
    url: 'https://your-tino-url'
paths:
  '/api/projects/{project_id}/servers/{server_id}/cron-jobs':
    get:
      summary: list
      operationId: list
      description: 'Get all cron jobs.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    -
                      id: null
                      server_id: null
                      command: 'ls -la'
                      user: root
                      frequency: '* * * * *'
                      status: ready
                      created_at: null
                      updated_at: null
                    -
                      id: null
                      server_id: null
                      command: 'ls -la'
                      user: root
                      frequency: '* * * * *'
                      status: ready
                      created_at: null
                      updated_at: null
                  links:
                    first: '/?page=1'
                    last: '/?page=1'
                    prev: null
                    next: null
                  meta:
                    current_page: 1
                    from: 1
                    last_page: 1
                    links:
                      -
                        url: null
                        label: '&laquo; Previous'
                        active: false
                      -
                        url: '/?page=1'
                        label: '1'
                        active: true
                      -
                        url: null
                        label: 'Next &raquo;'
                        active: false
                    path: /
                    per_page: 25
                    to: 2
                    total: 2
                properties:
                  data:
                    type: array
                    example:
                      -
                        id: null
                        server_id: null
                        command: 'ls -la'
                        user: root
                        frequency: '* * * * *'
                        status: ready
                        created_at: null
                        updated_at: null
                      -
                        id: null
                        server_id: null
                        command: 'ls -la'
                        user: root
                        frequency: '* * * * *'
                        status: ready
                        created_at: null
                        updated_at: null
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: null
                        server_id:
                          type: string
                          example: null
                        command:
                          type: string
                          example: 'ls -la'
                        user:
                          type: string
                          example: root
                        frequency:
                          type: string
                          example: '* * * * *'
                        status:
                          type: string
                          example: ready
                        created_at:
                          type: string
                          example: null
                        updated_at:
                          type: string
                          example: null
                  links:
                    type: object
                    properties:
                      first:
                        type: string
                        example: '/?page=1'
                      last:
                        type: string
                        example: '/?page=1'
                      prev:
                        type: string
                        example: null
                      next:
                        type: string
                        example: null
                  meta:
                    type: object
                    properties:
                      current_page:
                        type: integer
                        example: 1
                      from:
                        type: integer
                        example: 1
                      last_page:
                        type: integer
                        example: 1
                      links:
                        type: array
                        example:
                          -
                            url: null
                            label: '&laquo; Previous'
                            active: false
                          -
                            url: '/?page=1'
                            label: '1'
                            active: true
                          -
                            url: null
                            label: 'Next &raquo;'
                            active: false
                        items:
                          type: object
                          properties:
                            url:
                              type: string
                              example: null
                            label:
                              type: string
                              example: '&laquo; Previous'
                            active:
                              type: boolean
                              example: false
                      path:
                        type: string
                        example: /
                      per_page:
                        type: integer
                        example: 25
                      to:
                        type: integer
                        example: 2
                      total:
                        type: integer
                        example: 2
      tags:
        - cron-jobs
    post:
      summary: create
      operationId: create
      description: 'Create a new cron job.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  id: null
                  server_id: null
                  command: 'ls -la'
                  user: root
                  frequency: '* * * * *'
                  status: ready
                  created_at: null
                  updated_at: null
                properties:
                  id:
                    type: string
                    example: null
                  server_id:
                    type: string
                    example: null
                  command:
                    type: string
                    example: 'ls -la'
                  user:
                    type: string
                    example: root
                  frequency:
                    type: string
                    example: '* * * * *'
                  status:
                    type: string
                    example: ready
                  created_at:
                    type: string
                    example: null
                  updated_at:
                    type: string
                    example: null
      tags:
        - cron-jobs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                command:
                  type: string
                  description: ''
                  example: qui
                  nullable: false
                user:
                  type: string
                  description: ''
                  example: root
                  nullable: false
                  enum:
                    - root
                    - tino
                frequency:
                  type: string
                  description: 'Frequency of the cron job.'
                  example: '* * * * *'
                  nullable: false
              required:
                - command
                - user
                - frequency
    parameters:
      -
        in: path
        name: project_id
        description: 'The ID of the project.'
        example: 1
        required: true
        schema:
          type: integer
      -
        in: path
        name: server_id
        description: 'The ID of the server.'
        example: 3
        required: true
        schema:
          type: integer
  '/api/projects/{project_id}/servers/{server_id}/cron-jobs/{cronJob_id}':
    get:
      summary: show
      operationId: show
      description: 'Get a cron job by ID.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  id: null
                  server_id: null
                  command: 'ls -la'
                  user: root
                  frequency: '* * * * *'
                  status: ready
                  created_at: null
                  updated_at: null
                properties:
                  id:
                    type: string
                    example: null
                  server_id:
                    type: string
                    example: null
                  command:
                    type: string
                    example: 'ls -la'
                  user:
                    type: string
                    example: root
                  frequency:
                    type: string
                    example: '* * * * *'
                  status:
                    type: string
                    example: ready
                  created_at:
                    type: string
                    example: null
                  updated_at:
                    type: string
                    example: null
      tags:
        - cron-jobs
    delete:
      summary: delete
      operationId: delete
      description: 'Delete cron job.'
      parameters: []
      responses:
        204:
          description: ''
      tags:
        - cron-jobs
    parameters:
      -
        in: path
        name: project_id
        description: 'The ID of the project.'
        example: 1
        required: true
        schema:
          type: integer
      -
        in: path
        name: server_id
        description: 'The ID of the server.'
        example: 3
        required: true
        schema:
          type: integer
      -
        in: path
        name: cronJob_id
        description: 'The ID of the cronJob.'
        example: 5
        required: true
        schema:
          type: integer
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting <a href="/settings/api-keys" target="_blank">here</a>'
security:
  -
    default: []